+++ Microsoft.Azure.Cosmos/src/Tracing/TraceData/ConsistencyConfig.cs
@@ -1,0 +1,52 @@
+﻿// ------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.  All rights reserved.
+// ------------------------------------------------------------
+
+namespace Microsoft.Azure.Cosmos.Tracing.TraceData
+{
+    using System;
+    using System.Collections.Generic;
+    using System.Globalization;
+
+    internal class ConsistencyConfig
+    {
+        public ConsistencyConfig(
+            ConsistencyLevel? consistencyLevel,
+            IReadOnlyList<string> preferredRegions)
+        {
+            this.ConsistencyLevel = consistencyLevel.GetValueOrDefault();
+            this.PreferredRegions = preferredRegions;
+            this.lazyString = new Lazy<string>(() => string.Format(CultureInfo.InvariantCulture,
+                                "(consistency: {0}, prgns:[{1}])",
+                                consistencyLevel.GetValueOrDefault(),
+                                ConsistencyConfig.PreferredRegionsInternal(preferredRegions)));
+            this.lazyJsonString = new Lazy<string>(() => Newtonsoft.Json.JsonConvert.SerializeObject(this));
+        }
+
+        public ConsistencyLevel ConsistencyLevel { get; }
+        public IReadOnlyList<string> PreferredRegions { get; }
+
+        private readonly Lazy<string> lazyString;
+        private readonly Lazy<string> lazyJsonString;
+
+        public override string ToString()
+        {
+            return this.lazyString.Value;
+        }
+
+        public string ToJsonString()
+        {
+            return this.lazyJsonString.Value;
+        }
+
+        private static string PreferredRegionsInternal(IReadOnlyList<string> applicationPreferredRegions)
+        {
+            if (applicationPreferredRegions == null)
+            {
+                return string.Empty;
+            }
+
+            return string.Join(", ", applicationPreferredRegions);
+        }
+    }
+}
